split script into multiple files
