+++ src/booleanParameters.js
@@ -1,0 +1,5 @@
+exports.booleanParameters = [
+  'force',
+  'overwrite',
+  'long-opt-names'
+]
