+++ src/defaultsList.js
@@ -1,0 +1,34 @@
+exports.defaultsList = [
+  {
+    name: 'command',
+    default: 'render_list'
+  },
+  {
+    name: 'load',
+    value: '16'
+  },
+  {
+    name: 'map',
+    default: 'default'
+  },
+  {
+    name: 'proj',
+    default: 'EPSG:4326'
+  },
+  {
+    name: 'shell',
+    default: 'sh'
+  },
+  {
+    name: 'socket',
+    default: '/run/renderd/renderd.sock'
+  },
+  {
+    name: 'threads',
+    default: '1'
+  },
+  {
+    name: 'tile-dir',
+    default: '/var/lib/mod_tile'
+  }
+]
