+++ src/getTileCoordinates.js
@@ -1,0 +1,22 @@
+exports.getTileCoordinates = (lat, lon, zoom) => {
+  let xTile = Math.floor((lon + 180) / 360 * Math.pow(2, zoom))
+  let yTile = Math.floor(
+    (1.0 - Math.log(Math.tan(Math.PI * lat / 180.0)) + 1.0 / Math.cos(Math.PI * lat / 180.0) / Math.PI) * Math.pow(2.0, zoom - 1.0)
+  )
+  if (xTile < 0.0) {
+    xTile = 0.0
+  }
+  if (xTile >= Math.pow(2.0, zoom)) {
+    xTile = Math.pow(2.0, zoom)
+  }
+  if (yTile < 0.0) {
+    yTile = 0.0
+  }
+  if (yTile >= Math.pow(2.0, zoom)) {
+    yTile = Math.pow(2.0, zoom) - 1.0
+  }
+  return {
+    X: xTile,
+    Y: yTile
+  }
+}
