+++ src/longArgumentName.js
@@ -1,0 +1,15 @@
+exports.longArgumentName = {
+  '-a ': '--all',
+  '-f ': '--force',
+  '-l ': '--max-load=',
+  '-m ': '--map=',
+  '-n ': '--num-threads=',
+  '-s ': '---socket=',
+  '-t ': '--tile-dir=',
+  '-x ': '--min-x=',
+  '-X ': '--max-x=',
+  '-y ': '--min-y=',
+  '-Y ': '--max-y=',
+  '-z ': '--min-zoom=',
+  '-Z ': '--max-zoom='
+}
