+++ src/optionalArguments.js
@@ -1,0 +1,28 @@
+exports.optionalArguments = [
+  {
+    name: 'force',
+    skipVal: false,
+    option: '-f',
+    noValue: true
+  },
+  {
+    name: 'load',
+    skipVal: '16',
+    option: '-l'
+  },
+  {
+    name: 'map',
+    skipVal: 'default',
+    option: '-m'
+  },
+  {
+    name: 'threads',
+    skipVal: '1',
+    option: '-n'
+  },
+  {
+    name: 'tile-dir',
+    skipVal: '/var/lib/mod_tile',
+    option: '-t'
+  }
+]
